Expand Taxonomic Vectors into a Data Frame

This function takes a vector of species identifiers and expands it into a data frame that includes taxonomic classification information for each species. Each species identifier is expected to follow the format 'family_genus_species'.

Usage

expand_taxa(data = NULL)

Arguments

data

A vector of species identifiers where each entry is a string with the format 'family_genus_species'. If not supplied, the function will throw an error.

Value

A data frame where each row corresponds to a unique species from the input data. The data frame includes columns for species catch, number of elements, weight, and taxonomic classification including catch group, kingdom, phylum, order, family, genus and species.

Details

The function first splits the species identifier into a list of species, replaces underscores with spaces, and counts the number of words to help with further classification. It then uses conditional logic to extract genus and species names, handle special cases, and clean up group names. The cleaned data is then used to fetch taxonomic classification from the GBIF database using the taxize package. The final data frame is constructed by joining the expanded species data with the taxonomic ranks.

Note

The function requires internet access to fetch data from the GBIF database. It also assumes that the input vector is properly formatted. If there are any formatting issues or the GBIF database does not recognize a species, the function may return unexpected results or throw an error.

Examples

if (FALSE) { # \dontrun{
species_vector <- c("lutjanidae_lutjanus_spp", "scaridae_spp", "acanthuridae_naso_hexacanthus")
expanded_data <- expand_taxa(species_vector)
} # }